+++ cart/component/login/main.js
@@ -1,7 +1,9 @@
 import './style.css'
 import templateLogin from './index.html?raw';
+import * as HistoryComponent from '../history/main';
 const IconSrc = '../../../assets/icons';
 const user = new Object();
+const sesion = localStorage.getItem('iniciosesion');
 user.email = 'admin@admin.com';
 user.password = 'admin';
 
@@ -29,6 +31,28 @@ function Login() {
 		}
 	};
 }
+
+if (sesion){
+	console.log('sesion: ', sesion)
+	if (sesion === 'false') {
+		Login()
+	} else {
+		if(localStorage.getItem('statusback', 'cart')){
+			console.log('sesion: ', sesion)
+			HistoryComponent.render(app,readMovieList());
+			// renderHistory();
+		} else {
+			console.log('sesion: ', sesion)
+			window.location = '../../../shop/index.html';
+		}
+	}
+} else {
+	console.log('sesion: ', sesion)
+	localStorage.setItem('iniciosesion', false);
+	Login()
+}
+
+
 
 export function render(dom){
     dom.innerHTML = templateLogin.replace('${IconSrc}', IconSrc);
